refactor: created a custom hook for the submitNewGift function
